BM_Gen

Description

Documentation

main :: IO ()

main makes all the necessary function calls to generate a beatmap using and can be called via the console with the following command !!! Note: Almost all of these functions have been deprecated as of the demo_branch's creation. Don't judge my abilities based on this !!!

>>> ghc bm_gen.hs -e "main" > "outputfilename.txt"

boundsCheck :: (Integer, Integer, Integer) -> Integer

boundsCheck ensures that an object is within the given ranges

boundsCheck coordinateValue dimensionMinimum dimensionMaximum

findImpulse :: Integral a => Double -> [a] -> [Impulse]

findImpulse processes the audio data all while keeping track of successful impulses This is recursively by taking a chunk of 1470 samples to analyze and then recursing through the rest of the list

findImpulse frameCounter sampleList

differenceFinder :: Integral a => Double -> [a] -> [Impulse]

differenceFinder goes through the given list and returns an impulse when the difference between the amplitude of two partitions of size 735 is detected

impulse is then called in order to process the impulse and generate the RB and Osu buttons

processImpulse :: Complex Double -> Impulse

processImpulse calls genCircle in order to generate an Osu button !!! Note: Because of time constraints and issues with the generation algorithm, we always return an RB with A as a placeholder

This issue has been resolved and this function deprecated as of the demo_branch's creation

genCircle :: Complex Double -> Osu

genCircle generates an Osu button by taking the mod value of the x and y window. Note that here is a +/- 50 for either extreme in order to keep the button fully constrained within the expected radius

data Impulse

Impulse is the main data type used in order to encapsulate the activation frame and both the RB and Osu style buttons

!!! Note: I think adding documentation to Show and Generic would involve adding code so those won't be documented. They simply allow Impulse, RB, and Osu to be used as arbitrary classes and be displayed

Constructors

Impulse

Fields

act_frame :: Integer

rb :: RB

osu :: Osu

newtype RB

RB is a single value data type that contains the character that holds the output Char. It's useless in this version and only left in because of time constraints

!!! Note: See data Impulse for more information regarding Generic RB and Show RB

Constructors

RB

Fields

button :: Char

data Osu

Osu contains the (x,y) coordinates used to represent the location of an Osu style button !!! Note: See data Impulse for more information regarding Generic Osu and Show Osu

Constructors

Osu

Fields

x :: Integer

y :: Integer