Main

Description

Documentation

main :: IO ()

main makes all the necessary function calls to generate a beatmap using (currently) a single audio channel and when called with the proper commands via the console (listed below)

>>> ghc bm_gen.hs -e "main" > "outputfilename.txt"

splitList :: [[s]] -> [s]

splitList Returns the first element in a list of lists

WAVESamples -> [WAVESample]

>>> splitList audioSamples

impulse :: Int -> [Int32] -> [(Int, Int32)]

Returns a list impulses

Two adjacent 735 subsamples are compared and if the difference between them passes an (currently arbitrarily chosen by me) threshold it adds the the frame in which it occurs and the magnitude of the difference

>>> impulse counter [DATA.WAVESample] -> [(Int, WAVESample)]

pOut :: Int -> [(Int, WAVESample)] -> String

Formats data in a such a way as to be useable by OFF Keyboard Hero's game_logic.py. Our current format is:

Note #n - Activation Frame: f
{'Up': u, 'Down': d, 'Left': l, 'Right': r}